+++ modules/exit.js
@@ -1,0 +1,1 @@
+export const exitFileManager = userName => console.log(`Thank you for using File Manager, ${userName}!`)
