+++ modules/start.js
@@ -1,0 +1,1 @@
+export const startFileManager = () => console.log('Please start the application with command "npm run start -- --username=your_username"')
